+++ x/profiles/client/rest/rest.go
@@ -22,7 +22,7 @@ func RegisterRoutes(cliCtx client.Context, r *mux.Router) {
 type SaveProfileReq struct {
 	BaseReq  rest.BaseReq   `json:"base_req"`
 	DTag     string         `json:"dtag"`
-	Username string         `json:"username,omitempty"`
+	Nickname string         `json:"nickname,omitempty"`
 	Bio      string         `json:"bio,omitempty"`
 	Pictures types.Pictures `json:"pictures,omitempty"`
 }